+++ examples/domain-adaptation/plot_otda_semi_supervised.py
@@ -4,6 +4,9 @@
 OTDA unsupervised vs semi-supervised setting
 ============================================
 
+.. note::
+    Example added in release: 0.1.9.
+
 This example introduces a semi supervised domain adaptation in a 2D setting.
 It explicit the problem of semi supervised domain adaptation and introduces
 some optimal transport approaches to solve it.