+++ examples/gromov/plot_fgw_solvers.py
@@ -4,6 +4,9 @@
 Comparison of Fused Gromov-Wasserstein solvers
 ==============================
 
+.. note::
+    Example added in release: 0.9.1.
+
 This example illustrates the computation of FGW for attributed graphs
 using 4 different solvers to estimate the distance based on Conditional
 Gradient [24], Sinkhorn projections [12, 51] and alternated Bregman